% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beam_propagation.R
\name{beam_propagation}
\alias{beam_propagation}
\title{Radar beam propagation}
\usage{
beam_propagation(range, elevation, R = 6371000, Rp = 4 * R/3)
}
\arguments{
\item{range}{vector with the distance to the radar en meters.}

\item{elevation}{vector of the same length as \code{range} containing elevation
angles in degrees.}

\item{R}{radius of Earth in meters.}

\item{Rp}{aproximation used.}
}
\value{
A data frame conteing 3 variables:
\describe{
\item{ht}{height above the radar in meters.}
\item{rh}{horizontal range in meters.}
\item{lea}{local elevation angle in degrees.}
}
}
\description{
Calculates the propagation of the radar beam using the 4R/3 approximation.
}
