% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_wind.R
\docType{data}
\name{radial_wind}
\alias{radial_wind}
\title{One volume of radial velocity}
\format{A \code{data.table} with 2076960 rows and 4 variables:
\describe{
\item{range}{distance to de radar in meters}
\item{vr}{radial velocity in m/s}
\item{azimuth}{azimuth angle in degrees}
\item{elevation}{elevation angle in degrees}
}}
\source{
https://radar.inta.gob.ar/
}
\usage{
radial_wind
}
\description{
A single volume of radial velocity measured on January 14th 2016 between
06:00:05 and 06:04:31 at Paraná, Entre Ríos.
}
\keyword{datasets}
