% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvad-package.R
\docType{package}
\name{rvad-package}
\alias{rvad}
\alias{rvad-package}
\title{rvad: Velocity Azimuth Display of Meteorological Radars}
\description{
This package implement an algorithm to process radial wind from Doppler radars
    obtainning the proper horizontal wind. The implementation is based on Browning and Wexler
    (1968) <doi: 10.1175/1520-0450(1968)007%3C0105:TDOKPO%3E2.0.CO;2>
    with the addition of several quality controls.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paocorrales/rvad}
  \item Report bugs at \url{https://github.com/paocorrales/rvad/issues}
}

}
\author{
\strong{Maintainer}: Paola Corrales \email{paola.corrales@cima.fcen.uba.ar}

Other contributors:
\itemize{
  \item Elio Campitelli \email{elio.campitelli@cima.fcen.uba.ar} [contributor]
}

}
\keyword{internal}
