% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAD.R
\name{vad_fit}
\alias{vad_fit}
\title{Calculates the horizontal of every range and and elevation using radial wind.}
\usage{
vad_fit(vr, azimuth, range, elev_ang, max_na = 0.2,
  max_consecutive_na = 30, r2_min = 0.8)
}
\arguments{
\item{vr}{a vector containing the viento radial}

\item{azimuth}{a vector of length = length(vr) containing the azimuthal angle of every
vr observation.}

\item{range}{a vector of length = length(vr) containing the range (in meters) asociate
to the observation.}

\item{elev_ang}{a vector of length = length(vr) with the elevation angle of every observation.}

\item{max_na}{maximum percentage of missing data in a single ring (defined as the date in every
range and elevation angle).}

\item{max_consecutive_na}{maximun angular gap for a single ring.}

\item{r2_min}{minimum r squared permitted in each fit.}
}
\value{
A data frame with class `rvad_vad` containing 7 variables:
\describe{
\item{height}{height above the radar in meters.}
\item{u}{zonal wind in m/s.}
\item{v}{meridional wind in m/s.}
\item{range}{distance to the radar in meters.}
\item{elevation}{elevation angle in degrees.}
\item{r2}{r squared of the fit.}
\item{rmse}{root mean squeared error calculated as the standar deviation of the residuals.}
}
}
\description{
The radial wind must not have aliasing. Also, noise and other artifacts are not good for the algorithm.
Te funcction can work with sigle volumens of data scanned in PPI (Plan Position Indicator) mode.
}
