% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAD.R
\name{vad_fit}
\alias{vad_fit}
\title{Velocity Azimuth Display}
\usage{
vad_fit(radial_wind, azimuth, range, elevation, max_na = 0.2,
  max_consecutive_na = 30, r2_min = 0.8, azimuth_origin = 90,
  azimuth_direction = c("cw", "ccw"))
}
\arguments{
\item{radial_wind}{a vector containing the radial wind.}

\item{azimuth}{a vector of length = length(vr) containing the azimuthal angle
of every vr observation in degrees clockwise from 12 o' clock.}

\item{range}{a vector of length = length(vr) containing the range (in meters)
asociate to the observation.}

\item{elevation}{a vector of length = length(vr) with the elevation angle of
every observation in degrees.}

\item{max_na}{maximum percentage of missing data in a single ring (defined as
the date in every range and elevation angle).}

\item{max_consecutive_na}{maximun angular gap for a single ring.}

\item{r2_min}{minimum r squared permitted in each fit.}

\item{azimuth_origin}{angle that represents the zero azimuth in degrees
counterclockwise from the x axis.}

\item{azimuth_directon}{direction of the azimuth angle.}
}
\value{
A data frame with class \code{rvad_vad} that has a \code{\link[=plot]{plot()}} method and contains
7 variables:
\describe{
\item{height}{height above the radar in meters.}
\item{u}{zonal wind in m/s.}
\item{v}{meridional wind in m/s.}
\item{range}{distance to the radar in meters.}
\item{elevation}{elevation angle in degrees.}
\item{r2}{r squared of the fit.}
\item{rmse}{root mean squeared error calculated as the standar deviation of
the residuals.}
}
}
\description{
Approximates the horizontal components of the wind from radial wind measured
by Doppler radar using the Velocity Azimuth Display method from Browning and
Wexler (1968).
}
\details{
The algorithm can work with sigle volume of data scanned in PPI (Plan Position
Indicator) mode. The radial wind must not have aliasing. Removing the noise
and other artifacts is desirable.

\code{vad_fit()} takes vectors of the same length with radial wind, azimuth angle,
range and elevation angle and computes a sinusoidal fit for each ring of data
(the observation for a particular range and elevation) before doing a simple
quality control.

First, it checks if the amount of missing data (must be explicit on the data
frame) is greater than \code{max_na}, by default a ring with more than 20% of missing
data is descarted. Second, rejects any ring with a gap greater than
\code{max_consecutive_na}. Following Matejka y Srivastava (1991) the default is
set as 30 degrees. After the fit, the algorithm rejects rings whose fit has
a \code{r2} less than \code{r2_min}. It is recommended to define this threshold
after exploring the result with \code{r2_min = 0}.

Rings that fail any of the above-mentioned checks return \code{NA}.
}
\examples{
VAD <- with(radial_wind, vad_fit(radial_wind, azimuth, range, elevation))
plot(VAD)

}
\seealso{
\code{\link[=vad_regrid]{vad_regrid()}} to sample the result into a regular grid.
}
